Move styles to index.styles
